Add files via upload
This is the mid-semester exam period at the UENR campus in Sunyani. We were asked to create a table and a contact file. All the code has been zipped into one folder named UEB3222122. To access this folder, please extract the zip file to obtain the HTML and CSS files. You can then run the files on your laptop to view the HTML page for the table located in the "table" folder and the contact information in the "contact" folder.